asoc: max98088: fix cache restore operation
Bug: 862023

Reviewed-on: http://git-master/r/58416

+++ sound/soc/codecs/max98088.c
@@ -1592,7 +1592,7 @@ static int max98088_dai2_set_fmt(struct snd_soc_dai *codec_dai,
 
 static void max98088_sync_cache(struct snd_soc_codec *codec)
 {
-       u16 *reg_cache = codec->reg_cache;
+       u8 *reg_cache = codec->reg_cache;
        int i;
 
        if (!codec->cache_sync)
